+++ smacc2_performance_tools/performance_tests/sm_coretest_transition_speed_1/include/sm_coretest_transition_speed_1/states/st_state_2.hpp
@@ -31,7 +31,7 @@ using smacc2::EvStateRequestFinish;
 using smacc2::Transition;
 
 // STATE MACHINE SHARED VARIABLES (used in this state)
-extern unsigned int _counter_;
+extern uint64_t _counter_;
 
 // STATE DECLARATION
 struct State2 : smacc2::SmaccState<State2, SmCoretestTransitionSpeed1>
